fix(placeholder): handle when placeholder parent is already resolved
When we resolve a placeholder with another we link the otherwise orphan
placeholder to the other parent.

That parent can already be resolved and so we can't add that orphan
placeholder to its list of children to be resolved. So we do it on the
spot since we know the resolved parent and what relation the placeholder
had with it.

If you understood this your head must hurt.

Here's the minimal use case it solves:

```buzz
object Node {
    Node another,

    fun something() > Node {
        return this.another.somethingelse();
    }

    fun somethingelse() > Node {
        return this.another;
    }
}
```