fix(jit): Scope called function context alloca into a block
When JIT code calls a function, it allocate a slot on the stack to hold its NativeCtx.
That alloca was outliving the function being called since it was allocated outside of it.
tests/bench/008-for.buzz was crashing because we were doing 1M function call in the same function and so stacking 1M alloca.

=> We put the alloca inside a BSTART/BEND so the alloca is released once the function is being called is done