feat: Cheney GC
https://en.wikipedia.org/wiki/Cheney%27s_algorithm

TypeRegistry stops using type string representation has map key,
instead we compute our own hash. Still, we don't use the ObjTypeDef
itself as a key because it's not constant: types contain pointers
to other types that might be Placeholders and change content when
resolved. That would throw of the map because, when it needs resizing
it copies everything to a new map and the hash gives something different
since the content of the key has changed.